master: do not use stats feature
