+++ crates/xash3d-master/config/default.toml
@@ -30,7 +30,6 @@
 # Time in seconds before next admin request is allowed after wrong password
 #admin = 10
 
-# NOTE: require build with `stats` feature flag
 [stat]
 # Interval in seconds, set zero to disable
 #interval = 0
